nav: requests=public bourse, my-requests=personal feed (#1615)
- (tabs)/requests = public requests feed (visible to all, no auth)
- (tabs)/my-requests = current user's own requests
- public-requests tab → redirect to /(tabs)/requests
- update sidebar/drawer items + breadcrumbs + typed router
